Add files via upload
Configuring static overload NAT with ISP network, the configuration will create a 
network setup where a single public IP address will be statically assigned to multiple 
internal private IP addresses. It will then allow all internal devices on the network 
to be able to access internet with the use of same public addresses. 
The LAB will be done by placing two 2811 Cisco router, one for the ISP provider 
and the other for the organization and 2960 Cisco switch in the organization. The 
ISP router will connect to the organization router and the organization will be 
connected to the switch to provide network to all LAN devices from the router. The 
project will be done using Cisco Packet Tracer.